+++ src/components/login.vue
@@ -8,10 +8,10 @@
 			</div>
 			<div class="input-container">
 				<span class="icon-container" id="password-icon"></span>
-				<input type="password" class="form-input" placeholder="Password" v-model="password" />
+				<input type="password" class="form-input" placeholder="Password" @keyup="onKeyup" v-model="password" />
 			</div>
 			<div class="input-container">
-				<button class="form-btn" v-on:click="authenticate()">{{ formButton }}</button>
+				<button class="form-btn" v-on:click="authenticate">{{ formButton }}</button>
 			</div>
 			<div class="form-message" v-bind:class="{ error: flash.error }">
 				{{flash.message}}
@@ -52,6 +52,11 @@
 					this.flash.message = result.message;
 					this.flash.error = result.error;
 				});
+			},
+			onKeyup(e) {
+				if (e.keyCode === 13) {
+					this.authenticate();
+				}
 			}
 		}
 	}