+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.12.0] - 2017-11-26
+### Added
+- Setting chase coefficient
+- Setting the activation for the trader
+- Default route set to dashboard
+
 ## [0.11.0] - 2017-11-26
 ### Added
 - Setting preferences and saving to the database